model optimization and cleanup
