--- ZoomDotNetFramework/Responses/ZRError.cs
@@ -1,39 +1,0 @@
-﻿// <copyright>
-// Copyright 2021 by Kingdom First Solutions
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-// </copyright>
-//
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using ZoomDotNetFramework.Entities;
-using ZoomDotNetFramework.Responses;
-
-
-namespace ZoomDotNetFramework.Responses
-{
-    public class ZRError
-    {
-        public int Code { get; set; }
-        public string Message { get; set; }
-        public List<ZRErrorDataObject> Data {get; set; }
-    }
-    public class ZRErrorDataObject
-    {
-        public string Field { get; set; }
-        public string Message { get; set; }
-    }
-}
