+++ ZoomDotNetFramework/Responses/ZoomBaseResponse.cs
@@ -14,14 +14,6 @@
 // limitations under the License.
 // </copyright>
 //
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using ZoomDotNetFramework.Entities;
-using ZoomDotNetFramework.Responses;
-
 
 namespace ZoomDotNetFramework.Responses
 {
