+++ ZoomDotNetFramework/ZoomDotNetFramework.csproj
@@ -71,14 +71,12 @@
     <Compile Include="Enums\RoomType.cs" />
     <Compile Include="Responses\RoomsResponse.cs" />
     <Compile Include="Responses\ZoomBaseResponse.cs" />
-    <Compile Include="Responses\ZRError.cs" />
+    <Compile Include="Responses\ZRBaseMeetingInfo.cs" />
     <Compile Include="Responses\ZRErrorResponse.cs" />
-    <Compile Include="Responses\ZRScheduleResult.cs" />
+    <Compile Include="Responses\ZRCancelMeetingRequest.cs" />
     <Compile Include="Responses\ZRScheduleResponse.cs" />
     <Compile Include="Responses\ZRScheduleRequest.cs" />
     <Compile Include="Responses\ZRRequestBodyBase.cs" />
-    <Compile Include="Responses\ZRListResult.cs" />
-    <Compile Include="Responses\ZRBaseResponseResult.cs" />
     <Compile Include="Responses\ZRBaseResponse.cs" />
     <Compile Include="Responses\ZRListResponse.cs" />
     <Compile Include="ZoomApi.cs" />
